Bug 1681726 - Move c bindings to a separate file. r=aosmond
This will isolate the unsafe and make building without them easier.

Differential Revision: https://phabricator.services.mozilla.com/D99350
